Refactor chat components and structure
- Updated import paths for ThreadList and ChatContent components to reflect new directory structure.
- Added ChatContent component to handle chat interactions, including creating threads and sending messages.
- Introduced ThreadList component for displaying chat threads with options to select and delete threads.
- Created EmptyChatView component to show when no active thread is selected, including input for new messages and quick actions.
- Implemented MessageView component to render individual messages with support for markdown and file attachments.
- Added EmptyInputCard and QuickAction components for enhanced user experience in the empty chat state.
- Enhanced file handling and input features in MessageInput component, including drag-and-drop support.
