+++ package.json
@@ -95,6 +95,7 @@
     "npm": "^11.4.1",
     "postcss": "^8.5.4",
     "prettier": "^3.5.3",
+    "shadcn": "^3.4.0",
     "storybook": "^9.0.4",
     "storybook-dark-mode": "^4.0.2",
     "tailwind-merge": "^3.3.0",
